rest: move config into its own module
